Prime field modulus and Montgomery constants made static (#1)
Previous implementation of `RandomField` stored `ConfigRef` that described field modulus and precomputed values for Montgomery representation. Moreover, it allowed having `RandomField` with no config at all.
This approach introduced a lot of unnecessary complexity, and wasn't actually useful in any meaningful way.

This PR introduces a static `FieldConfig` that is passed to `RandomField` as a parameter. This also allowed us to get rid of lifetimes in  `RandomField` definition.

The resulting code is shorter, simpler and more performant.