Boolean evalutaions (#15)
(This is mostly a preparation PR to using binary decomposed polynomials)
* Switched Zip structs and traits to use semirings instead of rings
where appropriate.
* RAA code has been split onto RAA and sign-flipping RAA. This is
because sign-flipping version require `Neg` trait to be implemented, so
it won't work on semirings.
* Zip+ benches changed to use `Boolean` instead of `i8` for evaluation
coefficients.
* Additionally, test types for polynomials were changed to match
benchmark types of Zip+. (Note that for int test types it's still more
convenient to use types that differ to benchmarks)