Introduce traits `InnerProduct` and `InnerProductWithConfig` (#27)
There's a loose abstraction in both the paper and the implementation of
evaluating a univariate polynomial on a vector of points instead of
denoting it as an inner product. This PR fixes that in the code and
unifies various kinds of inner products by introducing the trait
`InnerProduct`.

Along the way, the dense univariate polys are moved into
`univariate::dense` module and the interface of the univariate
evaluation is slightly changed. Now, it doesn't expect a slice for
evaluation which typically was a slice of powers but rather a single
point whose powers it computes under the hood.

---------

Co-authored-by: Alex Abdugafarov <frozenspider@users.noreply.github.com>