Introduce binary polynomials
A binary polynomial is represeneted as `u32` (degree at most 31)
or `u64` (degree at most 63) padded to occupy a certain amount of bytes
(8, 16, or 32 bytes) to be able to turn into a big integer in place.

The conversion from a number to its bit polynomial is super cheap.

The projection function for binary polynomials benefits from
preprocessing the powers of the random challenge as then the evaluation
becomes just a filtered sum.

Added a couple benchmarks to compare ordinary (`DensePolynomial`) and
the newly inroduced binary polynomials.