Test phase optimisation (#40)
* Remove `.map(Zt::Comb::from_ref)` in
```rust
let evals: Vec<_> = poly
                .evaluations
                .iter()
                .map(Zt::Comb::from_ref)
                .map(|p| p.inner_product(&alphas, Zt::CombR::ZERO))
                .try_collect()?;
```
that regresses performance of the test phase. In the main benchmark,
`Zt::Comb` is `DensePolynomial<Int<5>, D_PLUS_ONE>` and `Zt::Eval` is
`DensePolynomial<Boolean, D_PLUS_ONE>`. So the conversion would erase
the information that the coefficients of the polynomial are 0s or 1s,
and a less optimal `inner_product` implementation would be used.

* Optimise trait constraints for `Eval` of `ZipTypes`. Now it's the bare
minimum of what we need to know about `Eval`.

---------

Co-authored-by: Alex Abdugafarov <frozenspider@users.noreply.github.com>