`InnerProduct` and `ProjectableToField` traits reworked (#44)
Currently, implementations of `InnerProduct` and `ProjectableToField`
can't be specialised for certain classes of types if a general
implentation provided. For instance, `DensePolynomial<Boolean, _>` can't
have its own specialised inner product if an implementation is provided
for `DensePolynomial<R, _> where R: Semiring`.

Inner product and projection to field shouldn't be tied to a (generic)
type, rather it should be possible to switch between different
implementations at will when it's necessary. It does clutter some
function calls but I believe it's worth having some freedom.

As a bonus: benchmarks of evaluation phase for zip+ are improved by 90%.

---------

Co-authored-by: Alex Abdugafarov <frozenspider@users.noreply.github.com>